moving DatasetPatch, DatasetStorage to components/dataset_storage.py
